backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.7
torch>=2.0.0
transformers>=4.35.0
Pillow>=9.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pydantic>=2.0.0